% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cacheDatabaseSets}
\alias{cacheDatabaseSets}
\title{getGSM2Target retrieves sample query sets.}
\usage{
cacheDatabaseSets(release = 2, dev = TRUE, verbose = TRUE)
}
\arguments{
\item{release}{Integer indicating the release number of the databaseSet
manifest to use. Optional. (Defualt: 2).}

\item{dev}{Logical value indiciating whether to use development version
of the manifest file. Optional. (Default: TRUE).}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
List of categorical vectors, each of which corresponding to a query
set.

One list of vectors corresponding to aggregated databaseSets.
}
\description{
getGSM2Target retrieves sample query sets.
}
\examples{
getGSM2Target('EPIC', 'TFBS')
cacheDatabaseSets cache databaseSets into memory

cacheDatabaseSets(release=2)
}
