% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clearCache}
\alias{clearCache}
\title{clearCache removes current cache (if it exists) otherwise returns an error.}
\usage{
clearCache(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\description{
clearCache removes current cache (if it exists) otherwise returns an error.
}
\examples{

}
