% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clearDatabaseSetCache}
\alias{clearDatabaseSetCache}
\title{clearDatabaseSetCache removes database set current cache (if it exists).}
\usage{
clearDatabaseSetCache(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\description{
clearDatabaseSetCache removes database set current cache (if it exists).
}
\examples{
clearDatabaseSetCache()

}
