% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clearFeatureEngineeringCache}
\alias{clearFeatureEngineeringCache}
\title{clearFeatureEngineeringCache removes universe set current cache (if it 
exists).}
\usage{
clearFeatureEngineeringCache(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\description{
clearFeatureEngineeringCache removes universe set current cache (if it 
exists).
}
\examples{
clearFeatureEngineeringCache()

}
