% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{createDatabaseSetNetwork}
\alias{createDatabaseSetNetwork}
\title{createGeneNetwork creates databaseSet network using the given similarity
metric.}
\usage{
createDatabaseSetNetwork(
  databaseSets,
  title = "Database Interaction Network",
  collection = "DatabaseSets"
)
}
\arguments{
\item{databaseSet}{Vector of probes corresponding to a single database set
of interest.}

\item{metric}{String representing the similarity score to use. Optional.
(Default: "Jaccard").}
}
\value{
ggplot lollipop plot
}
\description{
createGeneNetwork creates databaseSet network using the given similarity
metric.
}
\examples{
createDatabaseSetNetwork(list(a=c("a", "b"), b=c("a", "e", "f"), c=c("q", "a")))

}
