% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{databaseSetGet}
\alias{databaseSetGet}
\title{databaseSetGet retrieves database sets from a meta data sheet by querying the array, group, and reference columns. The data is returned as a list where the names correspond to chosen database sets.}
\usage{
databaseSetGet(keys = -1, platform = NA)
}
\arguments{
\item{keys}{vector containing the keys associated with the selected
databaseSets; only non-NA locations will be returned.}

\item{platform}{string representing the platform (EPIC, HM450, HM27, MM285)
for which database sets will be returned.}
}
\value{
One list containing features corresponding the test estimate, p-value, and type of test.
}
\description{
databaseSetGet retrieves database sets from a meta data sheet by querying the array, group, and reference columns. The data is returned as a list where the names correspond to chosen database sets.
}
\examples{
databaseSetGet(c(1))

}
