% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_master.R
\docType{data}
\name{df_master}
\alias{df_master}
\title{Master data frame for all object to cache}
\format{
A data.frame with columns:
\describe{
  \item{name}{Character. Resource name/title.}
  \item{eh}{Character. ExperimentHub record ID.}
}
}
\value{
master sheet of knowYourCG objects
}
\description{
This is an internal object which will be updated on every new release
library(ExperimentHub)
eh <- query(ExperimentHub(localHub=FALSE), "knowYourCG")
eh <- query(ExperimentHub(localHub=FALSE), "sesameData") # older data
data.frame(name=eh$title, eh=names(eh))
}
\details{
Cache location is default to
/Users/zhouw3/Library/Caches/org.R-project.R/R/ExperimentHub/
}
