% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getBetas}
\alias{getBetas}
\title{getBetas retrieves an example betas matrix composed of 20 samples.}
\usage{
getBetas(array, release = 1, dev = TRUE, verbose = TRUE)
}
\arguments{
\item{array}{string representing the array (EPIC, HM450, HM27, MM285)
for which database sets will be returned.}

\item{release}{Integer indicating the release number of the database set
manifest to use. Optional. (Defualt: 2).}

\item{dev}{Logical value indiciating whether to use development version
of the manifest file. Optional. (Default: TRUE).}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
betas matrix
}
\description{
getBetas retrieves an example betas matrix composed of 20 samples.
}
\examples{
getBetas("MM285")

}
