% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{getCGAnnotation}
\alias{getCGAnnotation}
\title{getCGAnnotation tests for the overlap of set of probes (querySet) in a
single given feature (database set)}
\usage{
getCGAnnotation(
  querySet,
  databaseSets = NA,
  release = NA,
  platform = NA,
  dev = TRUE
)
}
\arguments{
\item{databaseSets}{List of vectors corresponding to the databaseSets of
interest with associated meta data as an attribute to each element. Optional.
(Default: NA)}

\item{databaseSet}{Vector of probes corresponding to a single database set
of interest.}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}
}
\value{
An upper triangular matrix containing a metric (Jaccard) comparing
the pairwise distances between database sets.
}
\description{
getCGAnnotation tests for the overlap of set of probes (querySet) in a
single given feature (database set)
}
\examples{
getCGAnnotation(querySet=c("cg29176188_TC21", "cg29176794_TC21"), release=2)

}
