% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{getDatabaseSetOverlap}
\alias{getDatabaseSetOverlap}
\title{getDatabaseSetOverlap tests for the overlap of set of probes (querySet) in a
single given feature (database set)}
\usage{
getDatabaseSetOverlap(
  querySet,
  databaseSets = NA,
  release = NA,
  array = NA,
  dev = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{querySet}{Vector of probes corresponding to a single database set
of interest.}

\item{databaseSets}{List of vectors corresponding to the database sets of
interest with associated meta data as an attribute to each element. Optional.
(Default: NA)}

\item{release}{Integer indicating the release number of the database set. 
Optional. (Default: NA).}

\item{array}{String corresponding to the type of array to use. Either
MM285, EPIC, HM450, or HM27. If it is not provided, it will be inferred
from the query set probeIDs (Default: NA).}

\item{dev}{Logical value indicating whether to use development version.}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}
}
\value{
A sparse data.frame containing all of the meta data from all database
sets.
}
\description{
getDatabaseSetOverlap tests for the overlap of set of probes (querySet) in a
single given feature (database set)
}
\examples{
querySet=c("cg29176188_TC21", "cg29176794_TC21")
getDatabaseSetOverlap(querySet, release=2)

}
