% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getGSM2Target}
\alias{getGSM2Target}
\title{getGSM2Target retrieves sample query sets.}
\usage{
getGSM2Target(platform, modality)
}
\value{
List of categorical vectors, each of which corresponding to a query
set.
}
\description{
getGSM2Target retrieves sample query sets.
}
\examples{
getGSM2Target('EPIC', 'TFBS')
}
