% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getProbeID2Gene}
\alias{getProbeID2Gene}
\title{getProbeID2Gene retrieves mapping of of genes to probeID for a specific
array.}
\usage{
getProbeID2Gene(array)
}
\arguments{
\item{array}{string representing the array (EPIC, HM450, HM27, MM285)
for which databaseSets will be returned.}
}
\value{
List with each probe mapped to its respective gene.?
}
\description{
getProbeID2Gene retrieves mapping of of genes to probeID for a specific
array.
}
\examples{
getProbeID2Gene('EPIC')
}
