% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getQuerySets}
\alias{getQuerySets}
\title{getQuerySets retrieves sample query sets.}
\usage{
getQuerySets()
}
\value{
List of categorical vectors, each of which corresponding to a query
set.
}
\description{
getQuerySets retrieves sample query sets.
}
\examples{
getQuerySets()

}
