% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getTestQuerySet}
\alias{getTestQuerySet}
\title{getTestQuerySet retrieves sample query sets.}
\usage{
getTestQuerySet()
}
\value{
List of categorical vectors, each of which corresponding to a query
set.
}
\description{
getTestQuerySet retrieves sample query sets.
}
\examples{
getTestQuerySet()

}
