% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getUniverseSet}
\alias{getUniverseSet}
\title{getUniverseSet retrieves universe set of a given array.}
\usage{
getUniverseSet(array, verbose = TRUE)
}
\arguments{
\item{array}{string representing the array (EPIC, HM450, HM27, MM285)
for which databaseSets will be returned.}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
Vector of strings corresponding to the ProbeIDs on a given array.
}
\description{
getUniverseSet retrieves universe set of a given array.
}
\examples{
getUniverseSet("MM285.mm10.manifest")

}
