% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{inferarrayFromProbeIDs}
\alias{inferarrayFromProbeIDs}
\title{inferarrayFromProbeIDs infers the Infinium MicroArray array using the
given probeIDs}
\usage{
inferarrayFromProbeIDs(probeIDs)
}
\arguments{
\item{probeIDs}{Vector of probes of interest (e.g., probes belonging to a
given array)}
}
\value{
String corresponding to the inferred array
}
\description{
inferarrayFromProbeIDs infers the Infinium MicroArray array using the
given probeIDs
}
