% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{inferPlatformFromProbeIDs}
\alias{inferPlatformFromProbeIDs}
\title{inferPlatformFromProbeIDs infers the Infinium MicroArray platform using the
given probeIDs}
\usage{
inferPlatformFromProbeIDs(probeIDs)
}
\arguments{
\item{probeIDs}{Vector of probes of interest (e.g., probes belonging to a
given platform)}
}
\value{
String corresponding to the inferred platform
}
\description{
inferPlatformFromProbeIDs infers the Infinium MicroArray platform using the
given probeIDs
}
