% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plotEnrichAll.R
\name{KYCG_plotEnrichAll}
\alias{KYCG_plotEnrichAll}
\title{plot enrichment test result}
\usage{
KYCG_plotEnrichAll(
  df,
  fdr_max = 25,
  n_label = 15,
  min_estimate = 0,
  short_label = TRUE
)
}
\arguments{
\item{df}{test enrichment result data frame}

\item{fdr_max}{maximum fdr for capping}

\item{n_label}{number of database to label}

\item{min_estimate}{minimum estimate}

\item{short_label}{use short label}
}
\value{
grid object
}
\description{
plot enrichment test result
}
\examples{
query <- getDBs("MM285.designGroup")[["PGCMeth"]]
res <- testEnrichment(query, platform="MM285")
KYCG_plotEnrichAll(res)

}
