% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KYCG_plotPointRange.R
\name{KYCG_plotPointRange}
\alias{KYCG_plotPointRange}
\title{Plot point range for a list of enrichment testing results
against the same set of databases}
\usage{
KYCG_plotPointRange(result_list)
}
\arguments{
\item{result_list}{a list of testEnrichment resultsx}
}
\value{
grid plot object
}
\description{
Plot point range for a list of enrichment testing results
against the same set of databases
}
\examples{

## pick some big TFBS-overlapping CpG groups
cg_lists <- getDBs("MM285.TFBS")
queries <- cg_lists[(sapply(cg_lists, length) > 40000)]
result_list <- lapply(queries, testEnrichment,
    "MM285.chromHMM", platform="MM285")
KYCG_plotPointRange(result_list)

}
