% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kycgDataCache.R
\name{kycgDataCache}
\alias{kycgDataCache}
\title{Cache KnowYourCG data}
\usage{
kycgDataCache(data_titles = NULL)
}
\arguments{
\item{data_titles}{data to cache, if not given will cache all}
}
\value{
TRUE
}
\description{
Cache KnowYourCG data
}
\examples{
kycgDataCache("KYCG.HM27.Mask.20220123")
## to cache all data: kycgDataCache()
}
