% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kycgDataGet.R
\name{kycgDataGet}
\alias{kycgDataGet}
\title{Get KnowYourCG data}
\usage{
kycgDataGet(title, verbose = FALSE)
}
\arguments{
\item{title}{title of the data}

\item{verbose}{whether to output ExperimentHub message}
}
\value{
data object
}
\description{
Get KnowYourCG data
}
\examples{

kycgDataCache("KYCG.MSA.CGI.20220904")
EPIC.1.SigDF <- kycgDataGet('KYCG.MSA.CGI.20220904')
}
