% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotEnrichAll}
\alias{plotEnrichAll}
\title{plot enrichment test result}
\usage{
plotEnrichAll(df, fdr_max = 25, n_label = 15, min_estimate = 0)
}
\arguments{
\item{df}{test enrichment result data frame}

\item{fdr_max}{maximum fdr for capping}

\item{n_label}{number of database to label}

\item{min_estimate}{minimum estimate}
}
\value{
grid object
}
\description{
plot enrichment test result
}
\examples{
query <- getDBs("MM285.designGroup")[["PGCMeth"]]
res <- testEnrichment(query, platform="MM285")
plotEnrichAll(res)

}
