% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMeta}
\alias{plotMeta}
\title{Plot meta gene or other meta genomic features}
\usage{
plotMeta(betas, platform = NULL)
}
\arguments{
\item{betas}{a named numeric vector or a matrix
(row: probes; column: samples)}

\item{platform}{if not given and x is a SigDF, will be inferred
the meta features}
}
\value{
a grid plot object
}
\description{
Plot meta gene or other meta genomic features
}
\examples{
library(sesame)
sdf <- sesameData::sesameDataGet("EPIC.1.SigDF")
plotMeta(getBetas(sdf))
}
