% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMetaEnrichment}
\alias{plotMetaEnrichment}
\title{Plot meta gene or other meta genomic features}
\usage{
plotMetaEnrichment(result_list)
}
\arguments{
\item{result_list}{one or a list of testEnrichment}
}
\value{
a grid plot object
}
\description{
Plot meta gene or other meta genomic features
}
\examples{
cg_lists <- getDBs("MM285.TFBS")
queries <- cg_lists[(sapply(cg_lists, length) > 40000)]
result_list <- lapply(queries, testEnrichment,
    "MM285.metagene", silent=TRUE, platform="MM285")

plotMetaEnrichment(result_list)
}
