% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSetEnrichment}
\alias{plotSetEnrichment}
\title{Plot Set Enrichment}
\usage{
plotSetEnrichment(result, n_sample = 1000, n_presence = 200)
}
\arguments{
\item{result}{result object as returned from an element of the list of
testEnrichmentSEA(..., prepPlot=TRUE)}

\item{n_sample}{number of CpGs to sample}

\item{n_presence}{number of overlap to sample for the plot}
}
\value{
grid object for plot
}
\description{
Plot Set Enrichment
}
\examples{
query <- getDBs("KYCG.MM285.designGroup")[["VMR"]]
db <- getDBs("MM285.seqContextN", "distToTSS")
res <- testEnrichmentSEA(query, db, prepPlot = TRUE)
plotSetEnrichment(res[[1]])

}
