% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plotVolcano creates a volcano plot of -log2(p.value) and log(estimate)
given data with fields estimate and p.value.}
\usage{
plotVolcano(data, title = NA, xlabel = NA, ylabel = NA)
}
\arguments{
\item{data}{DataFrame where each field is a database name with two fields
for the estimate and p.value.}

\item{title}{String representing the title label. Optional. (Default: NA)}

\item{xlabel}{String representing the x-axis label. Optional. (Default:
NA)}

\item{ylabel}{String representing the y-axis label. Optional. (Default:
NA)}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
plotVolcano creates a volcano plot of -log2(p.value) and log(estimate)
given data with fields estimate and p.value.
}
