% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichment1}
\alias{testEnrichment1}
\title{testEnrichment1 tests for the enrichment of set of probes (query set) in a
single given feature (database set)}
\usage{
testEnrichment1(
  querySet,
  databaseSet,
  universeSet,
  estimate.type = "ES",
  p.value.adj = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{querySet}{Vector of probes of interest (e.g., significant probes)}

\item{universeSet}{Vector of probes in the universe set containing all of
the probes to be considered in the test.}

\item{estimate.type}{String indicating the estimate to report. (Default:
"ES")}

\item{p.value.adj}{Logical value indicating whether to report the adjusted
p-value. (Default: FALSE)}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}

\item{databaseSets}{List of vectors corresponding to the databaseSets of
interest with associated meta data as an attribute to each element.}
}
\value{
One list containing features corresponding the test estimate,
p-value, and type of test.
}
\description{
testEnrichment1 tests for the enrichment of set of probes (query set) in a
single given feature (database set)
}
