% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichmentAll}
\alias{testEnrichmentAll}
\title{testEnrichmentAll tests for the enrichment of set of probes (query set) in
a number of features (database sets).}
\usage{
testEnrichmentAll(
  querySet,
  databaseSets = NA,
  universeSet = NA,
  platform = NA,
  estimate.type = "ES",
  p.value.adj = FALSE,
  release = 2,
  verbose = FALSE
)
}
\arguments{
\item{querySet}{Vector of probes of interest (e.g., significant probes)}

\item{databaseSets}{List of vectors corresponding to the databaseSets of
interest with associated meta data as an attribute to each element. Optional.
(Default: NA)}

\item{universeSet}{Vector of probes in the universe set containing all of
the probes to be considered in the test. If it is not provided, it will be
inferred from the provided platform. (Default: NA)}

\item{platform}{String corresponding to the type of platform to use. Either
MM285, EPIC, HM450, or HM27. If it is not provided, it will be inferred
from the query set probeIDs (Default: NA)}

\item{estimate.type}{String indicating the estimate to report. (Default:
"ES")}

\item{p.value.adj}{Logical value indicating whether to report the adjusted
p-value. (Default: FALSE)}

\item{verbose}{Logical value indicating whether to display intermediate
text output about the type of test. Optional. (Default: FALSE)}
}
\value{
One list containing features corresponding the test estimate,
p-value, and type of test.
}
\description{
testEnrichmentAll tests for the enrichment of set of probes (query set) in
a number of features (database sets).
}
\examples{
testEnrichmentAll(c("cg0000029"))

}
