% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichmentFisher}
\alias{testEnrichmentFisher}
\title{testEnrichmentFisher uses Fisher's exact test to estimate the association
between two categorical variables.}
\usage{
testEnrichmentFisher(querySet, databaseSet, universeSet)
}
\arguments{
\item{querySet}{Vector of probes of interest (e.g., significant probes)}

\item{databaseSet}{Vector of probes corresponding to a single database set
of interest.}

\item{universeSet}{Vector of probes in the universe set containing all of
the probes to be considered in the test. (Default: NULL)}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
testEnrichmentFisher uses Fisher's exact test to estimate the association
between two categorical variables.
}
