% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGO.R
\name{testGO}
\alias{testGO}
\title{tests Gene Ontology of genes overlapping CpG query}
\usage{
testGO(query, platform = NULL, organism = "hsapiens", gene_name = TRUE, ...)
}
\arguments{
\item{query}{Vector of CpG probes IDs or a data frame with gene_name column,
usually the output of testEnrichment() function}

\item{platform}{EPIC, MM285, ..., infer if not given}

\item{organism}{The organism corresponding to the CpG platform
or genes in gene_name column}

\item{gene_name}{If query is data frame from testEnrichment output,
whether to use the gene_name column. If set to FALSE,
TFBS will be used (default: FALSE)}

\item{...}{Additional arguments to sesameData_getGenesByProbes and gost()}
}
\value{
A list of enriched terms and meta data from gprofiler2 output
}
\description{
estimate represent enrichment score and negative estimate indicate a
test for depletion
}
\examples{
library(SummarizedExperiment)
df <- rowData(sesameDataGet('MM285.tissueSignature'))
query <- df$Probe_ID[df$branch == "fetal_liver" & df$type == "Hypo"]
res <- testGO(query)
}
