% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{buildStatisticDataSet}
\alias{buildStatisticDataSet}
\title{buildStatisticDataSet builds dataset for each filename in filenames composed
of engineered features from the given databaseSets}
\usage{
buildStatisticDataSet(filenames, databaseSets)
}
\arguments{
\item{filenames}{Vector of filenames to load using readRDS/tbk_data}

\item{databaseSets}{List of vectors corresponding to probe locations for
which the features will be extracted}
}
\value{
DataFrame where rows are sample names and columns are features across
different databaseSets
}
\description{
buildStatisticDataSet builds dataset for each filename in filenames composed
of engineered features from the given databaseSets
}
