% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cacheDatabaseSets}
\alias{cacheDatabaseSets}
\title{cacheDatabaseSets cache databaseSets into memory}
\usage{
cacheDatabaseSets(locations, cacheLoc = "")
}
\arguments{
\item{locations}{Vector of strings corresponding to the URL locations of
databaseSets stored in RDS format.}

\item{cacheLoc}{String corresponding to the local filesystem location of
where the cache should be stored.}
}
\value{
One list of vectors corresponding to aggregated databaseSets.
}
\description{
cacheDatabaseSets cache databaseSets into memory
}
\examples{
cacheDatabaseSets(c("https://zhouserver.research.chop.edu/kyCG/20210713_EPIC_hg19_SNPs.rds", "https://zhouserver.research.chop.edu/kyCG/20210630_MM285_mm10_CTCF.rds"))
}
