% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cacheDatabaseSets}
\alias{cacheDatabaseSets}
\title{cacheDatabaseSets cache database sets into memory}
\usage{
cacheDatabaseSets(release = 2, dev = TRUE, verbose = TRUE)
}
\arguments{
\item{release}{Integer indicating the release number of the database set
manifest to use. Optional. (Defualt: 2).}

\item{dev}{Logical value indiciating whether to use development version
of the manifest file. Optional. (Default: TRUE).}

\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\value{
One list of vectors corresponding to aggregated database sets.
}
\description{
cacheDatabaseSets cache database sets into memory
}
\examples{
cacheDatabaseSets(release=2)
}
