% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clearCache}
\alias{clearCache}
\title{clearCache removes all current cache (if it exists).}
\usage{
clearCache(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value indicating whether intermediate outputs will be
displayed to console. Optional. (Default: TRUE).}
}
\description{
clearCache removes all current cache (if it exists).
}
\examples{
clearCache()

}
