% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getProbeID2Gene}
\alias{getProbeID2Gene}
\title{getProbeID2Gene retrieves mapping of of genes to probeID for a specific
platform.}
\usage{
getProbeID2Gene(platform)
}
\value{
List with each probe mapped to its respective gene.?
}
\description{
getProbeID2Gene retrieves mapping of of genes to probeID for a specific
platform.
}
\examples{
getProbeID2Gene('EPIC')
}
