% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{listDBGroups}
\alias{listDBGroups}
\title{List database group names}
\usage{
listDBGroups(filter = NULL, path = NULL, type = NULL)
}
\arguments{
\item{filter}{keywords for filtering}

\item{path}{file path to downloaded knowledgebase sets}

\item{type}{categorical, numerical (default: all)}
}
\value{
a list of db group names
}
\description{
List database group names
}
\examples{
head(listDBGroups("chromHMM"))
## or listDBGroups(path = "~/Downloads")
}
