% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{testEnrichment2}
\alias{testEnrichment2}
\title{Test enrichment from YAME-compressed CG sets
Concatenate Two Strings}
\usage{
testEnrichment2(
  query_fn,
  knowledge_fn,
  universe_fn = NULL,
  alternative = "greater"
)
}
\arguments{
\item{query_fn}{File path to query}

\item{knowledge_fn}{File path to knowledgebase}
}
\value{
A single concatenated string.
}
\description{
This function concatenates two strings using a C implementation.
}
\examples{
kfn = system.file("extdata", "chromhmm.cm", package = "knowYourCG")
qfn = system.file("extdata", "onecell.cg", package = "knowYourCG")
testEnrichment2(qfn, kfn)
}
