% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichmentAll}
\alias{testEnrichmentAll}
\title{Test for enriched categories of CpGs
testEnrichmentAll tests a default set of categories (database sets) for enrichment
in the provided vector of CpGs (sigProbes).}
\usage{
testEnrichmentAll(sigProbes, sigProbesRank = NULL, databaseSets = NULL)
}
\arguments{
\item{sigProbes}{Vector of probes of interested (e.g., significant probes)}

\item{sigProbesRank}{Numerical ranking for sigProbes such as P-values. Optional. (Default: NULL)}

\item{databaseSets}{Vector of database sets of interest. Optional. (Default: NULL)}
}
\value{
A list of two tables, one each for database sets of categorical vs continuous
values. The tables rank database sets by odds ratio (Fisher's exact test,
categorical sigProbes and categorical database set), or P-value (FGSEA or Spearman, categorical vs
continuous or continuous vs continuous respectively).
}
\description{
Test for enriched categories of CpGs
testEnrichmentAll tests a default set of categories (database sets) for enrichment
in the provided vector of CpGs (sigProbes).
}
