% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCorrelation.R
\name{testEnrichmentSpearman}
\alias{testEnrichmentSpearman}
\title{testEnrichmentSpearman uses the Spearman statistical test to estimate
the association between two continuous variables.}
\usage{
testEnrichmentSpearman(num_query, num_db)
}
\arguments{
\item{num_query}{named numeric vector of probes of interest 
where names are probe IDs (e.g significant probes)}

\item{num_db}{List of vectors corresponding to the database set of
interest with associated meta data as an attribute to each element.}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name
of test for the given results.
}
\description{
testEnrichmentSpearman uses the Spearman statistical test to estimate
the association between two continuous variables.
}
