% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{testEnrichmentSpearman}
\alias{testEnrichmentSpearman}
\title{testEnrichmentSpearman uses the spearman test to estimate the association
between two continuous variables.}
\usage{
testEnrichmentSpearman(querySet, databaseSet)
}
\arguments{
\item{querySet}{Vector of probes of interest (e.g., significant probes)}

\item{databaseSets}{List of vectors corresponding to the databaseSets of
interest with associated meta data as an attribute to each element.}
}
\value{
A DataFrame with the estimate/statistic, p-value, and name of test
for the given results.
}
\description{
testEnrichmentSpearman uses the spearman test to estimate the association
between two continuous variables.
}
